+++ src/main/java/org/nkjmlab/sorm4j/typed/TypedOrm.java
@@ -1,6 +1,5 @@
 package org.nkjmlab.sorm4j.typed;
 
-import org.nkjmlab.sorm4j.Orm;
 import org.nkjmlab.sorm4j.OrmMapReader;
 import org.nkjmlab.sorm4j.SqlExecutor;
 import org.nkjmlab.sorm4j.annotation.Experimental;
@@ -16,20 +15,5 @@ import org.nkjmlab.sorm4j.annotation.Experimental;
 public interface TypedOrm<T> extends TypedOrmReader<T>, TypedOrmUpdater<T>, OrmMapReader,
     TypedMetaDataFunction<T>, SqlExecutor {
 
-  /**
-   * Creates {@link TypedOrm}
-   *
-   * @param <S>
-   * @param objectClass
-   * @return
-   */
-  <S> TypedOrm<S> type(Class<S> objectClass);
-
-  /**
-   * Creates an {@link Orm}
-   *
-   * @return
-   */
-  Orm untype();
 
 }
